refactor: modify the logic to check first the deletion from storage

+++ trends/src/main/kotlin/net/thechance/trends/service/ReelsService.kt
@@ -43,8 +43,7 @@ class ReelsService(
             ?: throw ReelNotFoundException()
 
         runCatching {
-            reelsRepository.deleteById(id)
-            videoStorageService.deleteVideo(reel.videoUrl)
+            if (videoStorageService.deleteVideo(reel.videoUrl)) reelsRepository.deleteById(id)
         }.onFailure {
             throw VideoDeleteFailedException()
         }
